Events store method fixed, auth in method

+++ app/Controllers/Http/EventsController.ts
@@ -1,21 +1,28 @@
 import type { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
 import Event from 'App/Models/Event'
 import CreateEventValidator from 'App/Validators/CreateEventValidator'
-import { logger } from 'Config/app'
 
 export default class EventsController {
-  public async index({ response }: HttpContextContract) {
+  public async index({ response, auth }: HttpContextContract) {
     try {
-      const events = await Event.all()
+      const user = auth.use('api').user
+      if (!user) {
+        return response.status(401).json({ message: 'User not authenticated' })
+      }
+      const events = await Event.query().where('user_id', user.id)
       return response.status(200).json(events)
     } catch (error) {
       return response.status(error.status).json({ message: error.message })
     }
   }
 
-  public async show({ params, response }: HttpContextContract) {
+  public async show({ params, response, auth }: HttpContextContract) {
     try {
-      const event = await Event.findOrFail(params.id)
+      const user = auth.use('api').user
+      if (!user) {
+        return response.status(401).json({ message: 'User not authenticated' })
+      }
+      const event = await Event.query().where('id', params.id).where('userId', user.id).first()
       return response.status(200).json(event)
     } catch (error) {
       return response.status(error.status).json({ message: error.message })
@@ -24,25 +31,32 @@ export default class EventsController {
 
   public async store({ request, response, auth }: HttpContextContract) {
     try {
-      logger.info('Request body:', request.body()) // Log the request body
       const payload = await request.validate(CreateEventValidator)
-      logger.info('Payload after validation:', payload) // Log the payload after validation
       const user = auth.use('api').user
       if (!user) {
         return response.status(401).json({ message: 'User not authenticated' })
       }
+      console.log(user.id)
+
       const event = await Event.create({ ...payload, userId: user.id })
       return response.status(201).json(event)
     } catch (error) {
-      logger.info('Error:', error) // Log the error
-      return response.status(error.status).json({ message: error.message })
+      const statusCode = error.status && typeof error.status === 'number' ? error.status : 500
+      return response.status(statusCode).json({ message: error.message })
     }
   }
 
-  public async update({ params, request, response }: HttpContextContract) {
+  public async update({ params, request, response, auth }: HttpContextContract) {
     try {
       const payload = await request.validate(CreateEventValidator)
-      const event = await Event.findOrFail(params.id)
+      const user = auth.use('api').user
+      if (!user) {
+        return response.status(401).json({ message: 'User not authenticated' })
+      }
+      const event = await Event.query().where('id', params.id).where('userId', user.id).first()
+      if (!event) {
+        return response.status(404).json({ message: 'Event not found' })
+      }
       event.merge(payload)
       await event.save()
       return response.status(200).json(event)
@@ -51,9 +65,16 @@ export default class EventsController {
     }
   }
 
-  public async destroy({ params, response }: HttpContextContract) {
+  public async destroy({ params, response, auth }: HttpContextContract) {
     try {
-      const event = await Event.findOrFail(params.id)
+      const user = auth.use('api').user
+      if (!user) {
+        return response.status(401).json({ message: 'User not authenticated' })
+      }
+      const event = await Event.query().where('id', params.id).where('userId', user.id).first()
+      if (!event) {
+        return response.status(404).json({ message: 'Event not found' })
+      }
       await event.delete()
       return response.status(204)
     } catch (error) {
@@ -61,10 +82,16 @@ export default class EventsController {
     }
   }
 
-  public async search({ request, response }: HttpContextContract) {
+  public async search({ request, response, auth }: HttpContextContract) {
     try {
       const { title } = request.all()
-      const events = await Event.query().where('title', 'like', `%${title}%`)
+      const user = auth.use('api').user
+      if (!user) {
+        return response.status(401).json({ message: 'User not authenticated' })
+      }
+      const events = await Event.query()
+        .where('title', 'like', `%${title}%`)
+        .where('userId', user.id)
       return response.status(200).json(events)
     } catch (error) {
       return response.status(error.status).json({ message: error.message })
@@ -74,24 +101,6 @@ export default class EventsController {
   public async getEventsByUser({ params, response }: HttpContextContract) {
     try {
       const events = await Event.query().where('user_id', params.id)
-      return response.status(200).json(events)
-    } catch (error) {
-      return response.status(error.status).json({ message: error.message })
-    }
-  }
-
-  public async getPastEvents({ response }: HttpContextContract) {
-    try {
-      const events = await Event.query().where('date', '<', new Date())
-      return response.status(200).json(events)
-    } catch (error) {
-      return response.status(error.status).json({ message: error.message })
-    }
-  }
-
-  public async getUpcomingEvents({ response }: HttpContextContract) {
-    try {
-      const events = await Event.query().where('date', '>', new Date())
       return response.status(200).json(events)
     } catch (error) {
       return response.status(error.status).json({ message: error.message })
